refactor: 범용성을 고려하여 디렉토리 구조 수정
