+++ src/js/app.js
@@ -1,4 +1,4 @@
-import RacingGameController from './RacingGameController.js';
+import RacingGameController from './controllers/RacingGameController.js';
 
 class App {
   constructor() {
