+++ src/js/constants/RacingGame/selector.js
