+++ src/js/constants/RacingGame/setting.js
