+++ src/js/controllers/RacingGameController.js
@@ -1,9 +1,10 @@
-import RacingGameModel from './models/RacingGameModel.js';
-import RacingGameView from './RacingGameView.js';
-import nameStringToArray from './utils/nameStringToArray.js';
-import { $ } from './utils/element-tools.js';
-import { SELECTOR } from './constants/selector.js';
-import { isCarNameValid, isRaceTimeValid } from './utils/validator.js';
+import RacingGameModel from '../models/RacingGameModel.js';
+import RacingGameView from '../views/RacingGameView.js';
+
+import nameStringToArray from '../utils/nameStringToArray.js';
+import { $ } from '../utils/element-tools.js';
+import { SELECTOR } from '../constants/RacingGame/selector.js';
+import { isCarNameValid, isRaceTimeValid } from '../utils/RacingGame/validator.js';
 
 export default class RacingGameController {
   #racingGameModel;
@@ -22,20 +23,11 @@ export default class RacingGameController {
   }
 
   bindDefaultEvent() {
-    $(SELECTOR.CAR_NAME_BUTTON).addEventListener(
-      'click',
-      this.handleCarNameInput.bind(this)
-    );
+    $(SELECTOR.CAR_NAME_BUTTON).addEventListener('click', this.handleCarNameInput.bind(this));
 
-    $(SELECTOR.RACE_TIME_BUTTON).addEventListener(
-      'click',
-      this.handleRaceTimeInput.bind(this)
-    );
+    $(SELECTOR.RACE_TIME_BUTTON).addEventListener('click', this.handleRaceTimeInput.bind(this));
 
-    $(SELECTOR.RETRY_BUTTON).addEventListener(
-      'click',
-      this.handleGameRetry.bind(this)
-    );
+    $(SELECTOR.RETRY_BUTTON).addEventListener('click', this.handleGameRetry.bind(this));
   }
 
   handleCarNameInput(event) {
